refactor hypergeometric solver
